added an exported function for smooth scrolling
